requirements_detect.txt
ultralytics
opencv-python
numpy
Pillow
